[cpp] Use uint32_t to represent field masks
In various places (mainly the field_mask() functions) we previously
represented field masks with uint16_t's even though the ObjHeader
allocates 24 bits for them. Now we use uint32_t instead.

I'm motivated to do this now because my upcoming change to move the
object header outside of objects results in one type requiring 17 bits
for its field mask (the field masks get one bit longer).

+++ cpp/core.h
@@ -75,7 +75,7 @@ class PasswdEntry {
   int pw_uid;
   int pw_gid;
 
-  static constexpr uint16_t field_mask() {
+  static constexpr uint32_t field_mask() {
     return maskbit(offsetof(PasswdEntry, pw_name));
   }
 };
@@ -202,7 +202,7 @@ class SignalSafe {
     return result;
   }
 
-  static constexpr uint16_t field_mask() {
+  static constexpr uint32_t field_mask() {
     return maskbit(offsetof(SignalSafe, pending_signals_)) |
            maskbit(offsetof(SignalSafe, empty_list_));
   }
